Rename ThankYouPage from index.tsx
